+++ src/exporters/blocks.rs
@@ -12,10 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use futures::stream;
-use futures::stream::StreamExt;
-use futures::stream::TryStreamExt;
-
 use crate::contexts::ContextRef;
 use crate::eth::BlockFetcher;
 use crate::exceptions::Result;
@@ -34,25 +30,15 @@ impl BlockExporter {
     }
 
     pub async fn export(&self) -> Result<()> {
-        let jobs = self.numbers.chunks(self.ctx.get_batch_size()).len();
-        stream::iter(0..jobs)
-            .map(Ok)
-            .try_for_each_concurrent(self.ctx.get_max_worker(), |job| async move {
-                let mut fetcher = BlockFetcher::create(&self.ctx);
-                let mut chunks = self.numbers.chunks(self.ctx.get_batch_size());
-
-                if let Some(chunk) = chunks.nth(job) {
-                    fetcher.push_batch(chunk.to_vec())?;
-                    let blocks = fetcher.fetch().await?;
-                    let mut tx_hashes = vec![];
-                    for block in blocks {
-                        for tx in block.transactions {
-                            tx_hashes.push(tx.hash);
-                        }
-                    }
-                }
-                Ok(())
-            })
-            .await
+        let mut fetcher = BlockFetcher::create(&self.ctx);
+        fetcher.push_batch(self.numbers.to_vec())?;
+        let blocks = fetcher.fetch().await?;
+        let mut tx_hashes = vec![];
+        for block in blocks {
+            for tx in block.transactions {
+                tx_hashes.push(tx.hash);
+            }
+        }
+        Ok(())
     }
 }
